fix(compiler): scope SuppressAll native reconcile to linked harnesses
Link under agent_emission=suppress must not remove native agents on unrelated
targets. Thread selective_harness_scope through SuppressAll removal, extract
native_agents tests, and share reconcile+compile sequencing via a lifecycle helper.
